Updated examples, utils, tests

+++ src/edu/nps/moves/dis7/utilities/stream/X3dCreateInterpolators.java
@@ -75,7 +75,7 @@ public class X3dCreateInterpolators {
             //Same with X,Y,Z to create a local coordiante system
             if (firstTimeStamp) {
 
-                firstLocalTimeStamp = localPdu.getTimestamp();
+                firstLocalTimeStamp = localPdu.getTimestamp().intValue();
                 firstLocalX = localEspdu.getEntityLocation().getX();
                 firstLocalY = localEspdu.getEntityLocation().getZ();
                 firstLocalZ = -1 * localEspdu.getEntityLocation().getY();
@@ -83,7 +83,7 @@ public class X3dCreateInterpolators {
                 firstTimeStamp = false;
             }
 
-            localTimeStamp = localPdu.getTimestamp();
+            localTimeStamp = localPdu.getTimestamp().intValue();
             localX = localEspdu.getEntityLocation().getX();
             localY = localEspdu.getEntityLocation().getZ();
             localZ = -1 * localEspdu.getEntityLocation().getY();
